+++ demisto_sdk/main.py
@@ -3,6 +3,9 @@ from demisto_sdk.core import DemistoSDK
 
 def main():
     # TODO: Typings and docstrings
-
     sdk = DemistoSDK()
     sdk.parse_args()
+
+
+if __name__ == '__main__':
+    main()